Break out renderTimeRuler.
