+++ dom/render-activity-view.js
@@ -9,20 +9,15 @@ var comparators = require('../comparators');
 // var pluck = require('lodash.pluck');
 // var axis = require('d3-axis');
 var time = require('d3-time');
-var timeFormat = require('d3-time-format').timeFormat;
-var curry = require('lodash.curry');
+var renderTimeRuler = require('./render-time-ruler');
 var probable = require('probable');
 
-const yearBlockColor1 = '#fff';
-// hsl(108, 10%, 90%)
-const yearBlockColor2 = `hsl(${probable.roll(360)}, 10%, 90%)`;
 const activityFillHueA = probable.roll(360);
 const activityFillHueB = activityFillHueA + 40 % 360;
 const activityColorA = `hsl(${activityFillHueA}, 50%, 50%, 0.3)`;
 const activityColorB = `hsl(${activityFillHueB}, 50%, 50%, 0.3)`;
 
 const groupLabelMaxScale = 4;
-const maxYearLabelFontSize = 8; // in em.
 const baseDayLength = 40;
 const baseGroupSpacing = baseDayLength * 2;
 
@@ -36,22 +31,10 @@ const dayInMS = 24 * 60 * 60 * 1000;
 const today = new Date();
 const yesterday = new Date(today.getTime() - dayInMS);
 
-var formatMillisecond = timeFormat('.%L'),
-  formatSecond = timeFormat(':%S'),
-  formatMinute = timeFormat('%I:%M'),
-  formatHour = timeFormat('%I %p'),
-  formatDay = timeFormat('%a %d'),
-  formatWeek = timeFormat('%b %d'),
-  formatMonth = timeFormat('%B'),
-  formatYear = timeFormat('%Y');
-
 var activityContainer = d3.select('#activity-container');
 var labelBoard = d3.select('#label-board');
 //var activityGroupRoot = d3.select('#activity-groups');
 var fixedXRoot = labelBoard.select('.fixed-x-labels');
-var fixedYRoot = labelBoard.select('.fixed-y-labels');
-var dateLabels = fixedYRoot.select('.date-labels');
-var yearLabels = fixedYRoot.select('.year-labels');
 var targetsCanvas = d3.select('#activities-targets-canvas');
 var aCtx = d3
   .select('#activities-canvas')
@@ -125,7 +108,7 @@ function RenderActivityView() {
       .scaleTime()
       .domain([latestActivityDate, earliestActivityDate])
       .range([0, totalDaysSpan * baseDayLength]);
-      //.range([0, graphWidth]);
+    //.range([0, graphWidth]);
     zoomedTimeScale = timeScale;
 
     setUpZoom(draw);
@@ -134,7 +117,13 @@ function RenderActivityView() {
     function draw() {
       aCtx.clearRect(0, 0, graphWidth, graphHeight);
 
-      renderTimeRulers({ timeScale, graphWidth, graphHeight });
+      renderTimeRuler({
+        timeScale: zoomedTimeScale,
+        graphWidth,
+        graphHeight,
+        currentTransform,
+        ctx: aCtx
+      });
       renderGroupRulers({
         activityGroupData,
         graphWidth
@@ -168,7 +157,7 @@ function RenderActivityView() {
 
 function renderActivityGroups({
   activityGroupData,
-  timeScale,
+  timeScale
   // graphWidth,
   // graphHeight
 }) {
@@ -271,104 +260,8 @@ function hasDeeds(project) {
   return project && project.deeds && project.deeds.length > 0;
 }
 
-function renderTimeRulers({ graphWidth, graphHeight }) {
-  zoomedTimeScale.range([0, graphWidth]);
-  // zoomedTimeScale.domain(zoomedTimeScale.invert(0), zoomedTimeScale.invert(graphWidth));
-  var tickDates = zoomedTimeScale.ticks(); //time.timeMonth.every(1));
-  // If we're zoomed out far enough that the ticks are years, don't show them.
-  // The year blocks take care of that.
-  if (tickDates.length < 1 || time.timeYear(tickDates[0]) >= tickDates[0]) {
-    tickDates = [];
-  }
-  var tickYears = zoomedTimeScale.ticks(time.timeYear.every(1));
-  var yearWidth = graphWidth;
-  if (tickYears.length > 1) {
-    yearWidth = zoomedTimeScale(tickYears[1]) - zoomedTimeScale(tickYears[0]);
-  }
-
-  tickYears.forEach(curry(drawYearBlock)(yearWidth));
-
-  // weekRuler.attr('transform', 'translate(300, 0)');
-  // yearRuler.attr('transform', 'translate(300, 0)');
-  aCtx.strokeStyle = '#666';
-  aCtx.lineWidth = 1;
-  aCtx.beginPath();
-  tickDates.forEach(drawDateTick);
-  aCtx.stroke();
-
-  var tickTexts = dateLabels.selectAll('text').data(tickDates);
-  tickTexts.exit().remove();
-  tickTexts
-    .enter()
-    .append('text')
-    .attr('y', graphHeight / 8)
-    .merge(tickTexts)
-    .attr('x', zoomedTimeScale)
-    .text(multiFormat);
-
-  var yearLabelFontSize = 2 * currentTransform.k;
-  if (yearLabelFontSize > maxYearLabelFontSize) {
-    yearLabelFontSize = maxYearLabelFontSize;
-  }
-  var yearTexts = yearLabels.selectAll('text').data(tickYears);
-  yearTexts.exit().remove();
-  yearTexts
-    .enter()
-    .append('text')
-    .attr('transform', 'rotate(-90)')
-    .attr('x', -graphHeight / 2) // Because we're rotated, this is the vertical position.
-    .merge(yearTexts)
-    .attr('y', getYearLabelHorizontalPosition)
-    .attr('font-size', `${yearLabelFontSize}em`)
-    .text(formatYear);
-
-  function drawDateTick(date) {
-    var x = zoomedTimeScale(date);
-    aCtx.moveTo(x, 0);
-    aCtx.lineTo(x, graphHeight / 8);
-  }
-
-  function drawYearBlock(blockWidth, date) {
-    aCtx.fillStyle = date.getFullYear() % 2 ? yearBlockColor1 : yearBlockColor2;
-    aCtx.fillRect(zoomedTimeScale(date), 0, blockWidth, graphHeight);
-  }
-
-  function getYearLabelHorizontalPosition(date) {
-    return zoomedTimeScale(date) - currentTransform.k * graphWidth / 18;
-  }
-  // console.log('start', zoomedTimeScale.invert(0));
-  // console.log('end', zoomedTimeScale.invert(graphWidth));
-
-  // var weekAxis = axis.axisBottom(zoomedTimeScale);
-  // weekAxis.ticks(time.timeWeek.every(1));
-  // weekAxis.tickSize(dateTickLength);
-
-  // var yearAxis = axis.axisBottom(zoomedTimeScale);
-  // yearAxis.ticks(time.timeYear.every(1));
-  // yearAxis.tickFormat(timeFormat('%Y'));
-  // yearAxis.tickSize(100);
-
-  // weekRuler.call(weekAxis);
-  // yearRuler.call(yearAxis);
-}
-
 // function getGroupRulerTransform(name, i) {
 //   return `translate(${groupHeight * i}, 0)`;
 // }
-
-// Don't format years; leave that up to the year blocks.
-function multiFormat(date) {
-  return (time.timeSecond(date) < date
-    ? formatMillisecond
-    : time.timeMinute(date) < date
-      ? formatSecond
-      : time.timeHour(date) < date
-        ? formatMinute
-        : time.timeDay(date) < date
-          ? formatHour
-          : time.timeMonth(date) < date
-            ? time.timeWeek(date) < date ? formatDay : formatWeek
-            : formatMonth)(date);
-}
 
 module.exports = RenderActivityView;
